+++ src/models/listed-product.go
@@ -7,7 +7,7 @@ type ListedProduct struct {
 	CreatedAt      time.Time  `json:"createdAt"`
 	UpdatedAt      time.Time  `json:"updatedAt"`
 	DeletedAt      *time.Time `json:"deletedAt,omitempty"`
-	PriceListed    uint       `gorm:"not null,default:0" json:"priceListed"`
+	PriceListed    float64    `gorm:"not null,default:0" json:"priceListed"`
 	QuantityListed uint       `gorm:"not null,default:0" json:"quantityListed"`
 	InvoiceID      uint       `json:"invoiceId"`
 	ProductID      uint       `json:"productId"`
